services/api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
requests
pydantic_settings
prettytable
pandas
catboost
redis
fastapi-cache2[redis]
sqlalchemy
databases
psycopg2-binary
asyncpg
